Remove redundant files
